Conflicts:
	src/plugins/qnx/blackberrydeployconfiguration.cpp

+++ src/plugins/coreplugin/id.cpp
@@ -73,7 +73,7 @@ public:
         : n(length), str(s)
     {
         if (!n)
-            length = n = strlen(s);
+            length = n = static_cast<int>(strlen(s));
         h = 0;
         while (length--) {
             h = (h << 4) + *s++;
